refactor: split fix_cycle terminal-provenance and deferred-capture helpers out
CI kept failing the first-party 1500-line maintainability guardrail on
fix_cycle.py: every bot round adds lines, so trimming to just under the
limit only buys a round or two of headroom.

Move two self-contained blocks into sibling modules and re-export them
from fix_cycle with the existing ``X as X`` convention, so callers,
tests, and monkeypatch targets keep addressing them through fix_cycle:

- fix_cycle_terminal_provenance.py — the terminal ``_GitPushResult``
  shaping (unpushed local HEAD provenance, provenance-unavailable
  marking, agent verdict-protocol failure result).
- fix_cycle_deferred_capture.py — the deferred-thread durable capture
  (#305): idempotency marker, rendered conversation, and the capture
  itself.

Code is moved verbatim; only imports that no longer had a local user
were dropped from fix_cycle. fix_cycle.py goes 1531 -> 1222 lines, both
new modules are well under the limit, and no first-party file is now
over the 1500 gate.

One test imported ``_review_thread_body_hash`` as a pass-through from
fix_cycle; it now imports it from its home in awf.runtime.pr_monitor.
